Implemented possible implications
This means, that a token type can imply another token type (or actually multiple types).
For example, NumericLiteral implies DecimalIntegerLiteral. If the Parser requires a NumericLiteral, but received a DecimalIntegerLiteral, he can check if NumericLiteral.Implies(DecimalIntegerLiteral), and if that's true, the parser effectively received a NumericLiteral and can continue working.